PDEV-107 rename IMU -> LightningIMU & fix pigeon support
